APIs/azure.com/datalake-store-account/2016-11-01/python-aiohttp-client/tests/test_trusted_id_providers_controller.py
# coding: utf-8

import pytest
import json
from aiohttp import web

from openapi_server.models.create_or_update_trusted_id_provider_parameters import CreateOrUpdateTrustedIdProviderParameters
from openapi_server.models.trusted_id_provider import TrustedIdProvider
from openapi_server.models.trusted_id_provider_list_result import TrustedIdProviderListResult
from openapi_server.models.update_trusted_id_provider_parameters import UpdateTrustedIdProviderParameters


pytestmark = pytest.mark.asyncio

async def test_trusted_id_providers_create_or_update(client):
    """Test case for trusted_id_providers_create_or_update

    
    """
    parameters = {"properties":{"idProvider":"idProvider"}}
    params = [('api-version', 'api_version_example')]
    headers = { 
        'Accept': 'application/json',
        'Content-Type': 'application/json',
        'Authorization': 'Bearer special-key',
    }
    response = await client.request(
        method='PUT',
        path='/subscriptions/{subscription_id}/resourceGroups/{resource_group_name}/providers/Microsoft.DataLakeStore/accounts/{account_name}/trustedIdProviders/{trusted_id_provider_name}'.format(subscription_id='subscription_id_example', resource_group_name='resource_group_name_example', account_name='account_name_example', trusted_id_provider_name='trusted_id_provider_name_example'),
        headers=headers,
        json=parameters,
        params=params,
        )
    assert response.status == 200, 'Response body is : ' + (await response.read()).decode('utf-8')


pytestmark = pytest.mark.asyncio

async def test_trusted_id_providers_delete(client):
    """Test case for trusted_id_providers_delete

    
    """
    params = [('api-version', 'api_version_example')]
    headers = { 
        'Authorization': 'Bearer special-key',
    }
    response = await client.request(
        method='DELETE',
        path='/subscriptions/{subscription_id}/resourceGroups/{resource_group_name}/providers/Microsoft.DataLakeStore/accounts/{account_name}/trustedIdProviders/{trusted_id_provider_name}'.format(subscription_id='subscription_id_example', resource_group_name='resource_group_name_example', account_name='account_name_example', trusted_id_provider_name='trusted_id_provider_name_example'),
        headers=headers,
        params=params,
        )
    assert response.status == 200, 'Response body is : ' + (await response.read()).decode('utf-8')


pytestmark = pytest.mark.asyncio

async def test_trusted_id_providers_get(client):
    """Test case for trusted_id_providers_get

    
    """
    params = [('api-version', 'api_version_example')]
    headers = { 
        'Accept': 'application/json',
        'Authorization': 'Bearer special-key',
    }
    response = await client.request(
        method='GET',
        path='/subscriptions/{subscription_id}/resourceGroups/{resource_group_name}/providers/Microsoft.DataLakeStore/accounts/{account_name}/trustedIdProviders/{trusted_id_provider_name}'.format(subscription_id='subscription_id_example', resource_group_name='resource_group_name_example', account_name='account_name_example', trusted_id_provider_name='trusted_id_provider_name_example'),
        headers=headers,
        params=params,
        )
    assert response.status == 200, 'Response body is : ' + (await response.read()).decode('utf-8')


pytestmark = pytest.mark.asyncio

async def test_trusted_id_providers_list_by_account(client):
    """Test case for trusted_id_providers_list_by_account

    
    """
    params = [('api-version', 'api_version_example')]
    headers = { 
        'Accept': 'application/json',
        'Authorization': 'Bearer special-key',
    }
    response = await client.request(
        method='GET',
        path='/subscriptions/{subscription_id}/resourceGroups/{resource_group_name}/providers/Microsoft.DataLakeStore/accounts/{account_name}/trustedIdProviders'.format(subscription_id='subscription_id_example', resource_group_name='resource_group_name_example', account_name='account_name_example'),
        headers=headers,
        params=params,
        )
    assert response.status == 200, 'Response body is : ' + (await response.read()).decode('utf-8')


pytestmark = pytest.mark.asyncio

async def test_trusted_id_providers_update(client):
    """Test case for trusted_id_providers_update

    
    """
    parameters = {"properties":{"idProvider":"idProvider"}}
    params = [('api-version', 'api_version_example')]
    headers = { 
        'Accept': 'application/json',
        'Content-Type': 'application/json',
        'Authorization': 'Bearer special-key',
    }
    response = await client.request(
        method='PATCH',
        path='/subscriptions/{subscription_id}/resourceGroups/{resource_group_name}/providers/Microsoft.DataLakeStore/accounts/{account_name}/trustedIdProviders/{trusted_id_provider_name}'.format(subscription_id='subscription_id_example', resource_group_name='resource_group_name_example', account_name='account_name_example', trusted_id_provider_name='trusted_id_provider_name_example'),
        headers=headers,
        json=parameters,
        params=params,
        )
    assert response.status == 200, 'Response body is : ' + (await response.read()).decode('utf-8')

